backend/requirements.txt
torch
transformers
faiss-cpu
streamlit
